requirements.txt
#devops
black
click==8.1.3 
pytest==7.1.3
pytest-cov==4.0.0
#pylint==2.15.3
#rust based linter
ruff==0.0.284
mysql-connector-python
pandas
csvkit

